requirements.txt
anyio==3.7.1
click==8.2.1
colorama==0.4.6
fastapi==0.103.1
greenlet==3.2.3
h11==0.16.0
idna==3.10
pydantic==1.10.13
PyMySQL==1.1.1
python-dotenv==1.1.0
python-multipart==0.0.20
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlmodel==0.0.24
starlette==0.27.0
typing_extensions==4.14.0
uvicorn==0.34.3
cryptography
passlib[bcrypt]
pyjwt
